k.join can move between vcs and will say smth if alr in the vc k.ge censors server names added permission checks changed cookies location updated on_command_error replaced os.removes with custom ver. removed unused test command "k.sticker" k.uploadsticker fix for servers w/ 5 free sticker slots may_sounds variable relocated

+++ modules/vc.py
@@ -8,8 +8,6 @@ import os
 import asyncio
 import json
 
-may_sounds = ["sounds/totsugeki_7UWR0L4.mp3", "sounds/totsugeki-may-2.mp3"]
-
 
 class Vc(commands.Cog):
     def __init__(self, client):
@@ -20,12 +18,21 @@ class Vc(commands.Cog):
         voicestate = ctx.author.voice
         if voicestate:
             voice_channel = ctx.author.voice.channel
+            can_connect = voice_channel.permissions_for(ctx.guild.me).connect
+            can_speak = voice_channel.permissions_for(ctx.guild.me).speak
 
         voice = disnake.utils.get(self.client.voice_clients, guild=ctx.guild)
         if voicestate != None:
-
             if voice == None:
-                voice = await voice_channel.connect()
+                if can_connect and can_speak:
+                    voice = await voice_channel.connect()
+                else:
+                    missing_perms = []
+                    if not can_connect:
+                        missing_perms.append("connect")
+                    if not can_speak:
+                        missing_perms.append("speak")
+                    raise commands.BotMissingPermissions(missing_perms)
             if loop == "loop":
 
                 def loop():
@@ -113,13 +120,41 @@ class Vc(commands.Cog):
         await ctx.message.delete()
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def join(self, ctx):
-        voice_channel = ctx.author.voice.channel
-        await voice_channel.connect()
-        await ctx.send("Sus bot has joined the call.", delete_after=3.0)
-        await ctx.message.delete()
+        voicestate = ctx.author.voice
+        if voicestate:
+            voice_channel = ctx.author.voice.channel
+        else:
+            await ctx.send("Brooo you're not even in a vc. Get in one bruh")
+            return
+        voice = disnake.utils.get(self.client.voice_clients, guild=ctx.guild)
+
+        can_connect = voice_channel.permissions_for(ctx.guild.me).connect
+        can_speak = voice_channel.permissions_for(ctx.guild.me).speak
+        if can_connect and can_speak:
+            if not voice:
+                await voice_channel.connect()
+                await ctx.send("Sus bot has joined the call.", delete_after=3.0)
+                await ctx.message.delete()
+            elif ctx.guild.me.voice.channel == voice_channel:
+                await ctx.send("I'm already in the VC! Zamn.", delete_after=3.0)
+                await ctx.message.delete()
+            else:
+                await ctx.guild.voice_client.disconnect()
+                await voice_channel.connect()
+                await ctx.send("Sus bot has moved to this VC.", delete_after=3.0)
+                await ctx.message.delete()
+        else:
+            missing_perms = []
+            if not can_connect:
+                missing_perms.append("connect")
+            if not can_speak:
+                missing_perms.append("speak")
+            raise commands.BotMissingPermissions(missing_perms)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def stop(self, ctx):
         voice_client = ctx.message.guild.voice_client
         if voice_client.is_playing():
@@ -132,6 +167,7 @@ class Vc(commands.Cog):
         await ctx.message.delete()
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def stoploop(self, ctx):
         await ctx.guild.voice_client.disconnect()
         voice_channel = ctx.author.voice.channel
@@ -140,24 +176,7 @@ class Vc(commands.Cog):
         await ctx.message.delete()
 
     @commands.command()
-    async def letsgo(self, ctx, loop=None):
-        await self.vcplay(ctx, "sounds/vibez-lets-go.mp3", loop)
-
-    @commands.command()
-    async def vtubus(self, ctx, loop=None):
-        await self.vcplay(ctx, "sounds/vtubus.mp3", loop)
-
-    @commands.command()
-    async def giorno(self, ctx, loop=None):
-        await self.vcplay(ctx, "sounds/giorno theme.mp3", loop)
-
-    @commands.command()
-    async def ding(self, ctx, loop=None):
-        await self.vcplay(
-            ctx, "sounds/DING DING DING DING DING DING DING DI DI DING.mp3", loop
-        )
-
-    @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def leave(self, ctx):
         if ctx.voice_client:  # If the bot is in a voice channel
             await ctx.guild.voice_client.disconnect()  # Leave the channel
@@ -170,38 +189,67 @@ class Vc(commands.Cog):
             )
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
+    async def letsgo(self, ctx, loop=None):
+        await self.vcplay(ctx, "sounds/vibez-lets-go.mp3", loop)
+
+    @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
+    async def vtubus(self, ctx, loop=None):
+        await self.vcplay(ctx, "sounds/vtubus.mp3", loop)
+
+    @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
+    async def giorno(self, ctx, loop=None):
+        await self.vcplay(ctx, "sounds/giorno theme.mp3", loop)
+
+    @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
+    async def ding(self, ctx, loop=None):
+        await self.vcplay(
+            ctx, "sounds/DING DING DING DING DING DING DING DI DI DING.mp3", loop
+        )
+
+    @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def yodayo(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/Nakiri Ayame's yo dayo_.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def yodazo(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/Yo Dazo!.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def jonathan(self, ctx, loop=None):
         await self.vcplay(
             ctx, "sounds/Jonathan's theme but its only the BEST part.mp3", loop
         )
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def joseph(self, ctx, loop=None):
         await self.vcplay(
             ctx, "sounds/Joseph's theme but only the good part (1).mp3", loop
         )
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def jotaro(self, ctx, loop=None):
         await self.vcplay(
             ctx, "sounds/Jotaro’s theme but it’s only the good part.mp3", loop
         )
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def josuke(self, ctx, loop=None):
         await self.vcplay(
             ctx, "sounds/Josuke theme but it's only the good part.mp3", loop
         )
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def kira(self, ctx, loop=None):
         await self.vcplay(
             ctx,
@@ -210,98 +258,122 @@ class Vc(commands.Cog):
         )
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def pillarmen(self, ctx, loop=None):
         await self.vcplay(
             ctx, "sounds/Jojo's Bizarre Adventure- Awaken(Pillar Men Theme).mp3", loop
         )
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def boom(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/boom.mp3", loop)
 
     @commands.command(aliases=["ogei"])
+    @commands.bot_has_permissions(manage_messages=True)
     async def ogey(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/ogey.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def rrat(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/rrat.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def fart(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/fart.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def mogumogu(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/mogu.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def bababooey(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/bababooey.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def dog(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/dog.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def totsugeki(self, ctx, loop=None):
+        may_sounds = ["sounds/totsugeki_7UWR0L4.mp3", "sounds/totsugeki-may-2.mp3"]
         await self.vcplay(ctx, may_sounds, loop, True)
 
     @commands.command(aliases=["bong"])
+    @commands.bot_has_permissions(manage_messages=True)
     async def tacobell(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/tacobell.mp3", loop)
 
     @commands.command(aliases=["amogus"])
+    @commands.bot_has_permissions(manage_messages=True)
     async def amongus(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/amongus.mp3", loop)
 
     @commands.command(aliases=["classtrial"])
+    @commands.bot_has_permissions(manage_messages=True)
     async def danganronpa(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/danganronpa.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def botansneeze(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/botansneeze.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def water(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/water.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def necoarc(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/necoarc.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def vsauce(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/vsauce.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def gigachad(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/gigachad.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def bruh(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/rushiabruh.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def believeit(self, ctx, loop=None):
         believeit_files = os.listdir("sounds/believeit/")
         believeit_files = [f"sounds/believeit/{x}" for x in believeit_files]
         await self.vcplay(ctx, believeit_files, loop, True)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_messages=True)
     async def pikamee(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/pikamee.mp3", loop)
 
     @commands.command(aliases=["hellskitchen", "violin"])
+    @commands.bot_has_permissions(manage_messages=True)
     async def waterphone(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/waterphone.mp3", loop)
 
     @commands.command(aliases=["boo-womp"])
+    @commands.bot_has_permissions(manage_messages=True)
     async def boowomp(self, ctx, loop=None):
         await self.vcplay(ctx, "sounds/boowomp.mp3", loop)
 
     @commands.command()
+    @commands.bot_has_permissions(manage_webhooks=True, manage_messages=True)
     async def mgr(self, ctx, *, name):
         loop = None
         matches = []
